tests/wpt/web-platform-tests/XMLHttpRequest/setrequestheader-allow-whitespace-in-value.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: setRequestHeader() - header value with whitespace</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-setrequestheader()-method" data-tested-assertations="/following::ol/li[4]/p[contains(@class,'note')] /following::ol/li[6]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      function request(value) {
        test(function() {
          var client = new XMLHttpRequest()
          client.open("POST", "resources/inspect-headers.py?filter_name=X-Empty", false)
          client.setRequestHeader('X-Empty', value)
          client.send(null)
          assert_equals(client.responseText, 'x-empty: '+ String(value.trim()).toLowerCase()+'\n' )
        }, document.title + " (" + value + ")")
      }
      request(" ")
      request(" t")
      request("t ")
      request(" t ")
    </script>
  </body>
</html>
